Update 1_🗃️_Survey_Components.py
I encountered a problem with the numerical input component being marked as "st." which prevented the creation of a JSON input, and I was unable to reach out again to the 60 participants involved in my initial project. I believe that instead of being labeled as "st.," it should have been designated as "survey." for clarity.